backend/pyproject.toml
[project]
name = "finance-tracker-api"
version = "4.0.0"
description = "Expense tracking API with receipt OCR, bank import, debt tracking, AI chat, and Telegram bot"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.7",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=10.3.0",
    "pytesseract>=0.3.10",
    "pdfplumber>=0.11.0",
    "pdf2image>=1.17.0",
    "pandas>=2.2.0",
    "rapidfuzz>=3.9.0",
    "anthropic>=0.39.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.1",
    "bcrypt>=4.0.0",
    "httpx>=0.27.0",
    "python-telegram-bot>=21.0",
    "aiohttp>=3.9.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.4.8",
    "aiosqlite>=0.22.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
]
